MediatR Unit-test and mock, UT Code refactor
