+++ Yummi.UnitTest/UT_RecipeIngredientController.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Yummi.UnitTest
+{
+    internal class UT_RecipeIngredientController
+    {
+    }
+}
